QAuthenticator: Don't assume cached credentials are new
Although this is more of a QNAM thing.
In some cases what we load from the cache might be credentials we have
already tried (and failed with, thus leading us to emit the
authenticationRequired signal). With this patch we will fall through
more often and ask the user for credentials.

Change-Id: If2a556883c3ea5b0b225f4df273d38353b552b54
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
(cherry picked from commit 9021d4bbf0795cae601fe2bd656e15ae63f97e61)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>