Android: Pass the parsed URI if it fails to get permission on 23 or less
In Android API level 23 or lower it will still be able to open the file
even if the permissions are not provided. If the system is not going to
allow it then it will prevent it later on. So we trust the system to
block access if need be.

Change-Id: If663e3cf0a10818931577988abdc0fb8ec2650db
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 37f4c6fbc28617fdd3343033c6f926ec2263fce2)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>