Revert "Do not wait in QWindowsPipe{Reader|Writer}::stop()"
This reverts commit c7ec07d40115bef849574c81d619b629af9434a9.

Reason for revert: This causes a memory leak on program termination.

The initial commit was an attempt to fix a deadlock where the user
destroys object from a thread that does not own the object. This is
an unsupported case and should be treated as an invalid report.

Change-Id: I4957784b86a0361adb65b9d023542f96480f00ba
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
(cherry picked from commit 92640829ad9f14ac83d870ed119e2738983f5284)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>