Do not expose ATSPI as public dependency of QtGui
This was always a compile-only dependency and atspi includes are only
used in private headers. The atspi headers can never be a requirement
for consumers of QtGui.

In Qt5 this code was part of the QtLinuxAccessibilitySupport private
module, and for that module the atspi dependency was public. This only
affected users that explicitly instructed qmake to pull in the
linuxaccessibility_support-private module.

Task-number: QTBUG-86421
Change-Id: I23c6a1c466f86b2015dde13fca605fd26299c9db
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>